+++ ChangeLog.rst
@@ -1,3 +1,18 @@
+.. cpp:namespace:: graph_canon
+.. default-role:: cpp:expr
+
+Changes
+#######
+
+Develop
+==========================
+
+- Require version v0.3 PermGroup.
+- Add Schreier-Sims-based automorphism pruning.
+- Documentation polishing.
+- The interface of `aut_pruner_base` has been changed.
+- `Visitor::canon_new_best` has been changed.
+- Running the algorithm on an empty graph now works.
 
 Release v0.2 (2018-02-05)
 ==========================